+++ src/Tickets/Commerce/Gateways/Contracts/Abstract_Webhooks.php
@@ -14,6 +14,15 @@ use Tribe__Utils__Array as Arr;
 abstract class Abstract_Webhooks {
 
 	/**
+	 * Option name for the option to store pending webhooks.
+	 *
+	 * @since TBD
+	 *
+	 * @var string
+	 */
+	public const PENDING_WEBHOOKS_KEY = '_tec_tickets_commerce_webhook_pending';
+
+	/**
 	 * Gets the gateway for this webhook.
 	 *
 	 * @since 5.3.0
@@ -105,4 +114,91 @@ abstract class Abstract_Webhooks {
 		return delete_option( $this->get_settings_key() );
 	}
 
+	/**
+	 * Add a pending webhook to the order.
+	 *
+	 * @since 5.18.1
+	 * @since TBD Moved to the Abstract_Webhooks class from the Stripe_Webhooks class.
+	 *
+	 * @param int    $order_id   Order ID.
+	 * @param string $new_status New status.
+	 * @param string $old_status Old status.
+	 * @param array  $metadata   Metadata.
+	 *
+	 * @return void
+	 */
+	public function add_pending_webhook( int $order_id, string $new_status, string $old_status, array $metadata = [] ): void {
+		add_post_meta(
+			$order_id,
+			self::PENDING_WEBHOOKS_KEY,
+			[
+				'new_status' => $new_status,
+				'metadata'   => $metadata,
+				'old_status' => $old_status,
+			]
+		);
+	}
+
+	/**
+	 * Get the pending webhooks for an order.
+	 *
+	 * @since 5.18.1
+	 * @since TBD Moved to the Abstract_Webhooks class from the Stripe_Webhooks class.
+	 *
+	 * @param int $order_id Order ID.
+	 *
+	 * @return array
+	 */
+	public function get_pending_webhooks( int $order_id ): array {
+		return (array) get_post_meta( $order_id, static::PENDING_WEBHOOKS_KEY );
+	}
+
+	/**
+	 * Delete the pending webhooks for an order.
+	 *
+	 * @since 5.18.1
+	 * @since TBD Moved to the Abstract_Webhooks class from the Stripe_Webhooks class.
+	 *
+	 * @param int $order_id Order ID.
+	 *
+	 * @return void
+	 */
+	public function delete_pending_webhooks( int $order_id ): void {
+		delete_post_meta( $order_id, static::PENDING_WEBHOOKS_KEY );
+	}
+
+	/**
+	 * Get the max number of retries for the webhooks.
+	 *
+	 * @since 5.19.3
+	 * @since TBD Moved to the Abstract_Webhooks class from the Stripe_Webhooks class. and added generic filter.
+	 *
+	 * @return int The number of retries.
+	 */
+	public function get_max_number_of_retries(): int {
+		/**
+		 * Filter the maximum number of attempts we will try to retry a webhook process.
+		 *
+		 * @since TBD
+		 *
+		 * @param int $max_attempts How many attempts we will try to retry a webhook process. Defaults to 5.
+		 *
+		 * @return int
+		 */
+		$max_retries = (int) apply_filters( 'tec_tickets_commerce_gateway_webhook_maximum_attempts', 5 );
+
+		$gateway     = $this->get_gateway();
+		$gateway_key = $gateway::get_key();
+
+		/**
+		 * Filter the maximum number of attempts we will try to retry a webhook process for a specific gateway.
+		 *
+		 * @since 5.19.3
+		 *
+		 * @param int $max_attempts How many attempts we will try to retry a webhook process. Defaults to 5.
+		 *
+		 * @return int
+		 */
+		return (int) apply_filters( "tec_tickets_commerce_gateway_{$gateway_key}_webhook_maximum_attempts", $max_retries );
+	}
 }
