+++ src/Tickets/Commerce/Gateways/Stripe/Webhooks.php
@@ -99,56 +99,6 @@ class Webhooks extends Abstract_Webhooks {
 	}
 
 	/**
-	 * Add a pending webhook to the order.
-	 *
-	 * @since 5.18.1
-	 *
-	 * @param int    $order_id   Order ID.
-	 * @param string $new_status New status.
-	 * @param string $old_status Old status.
-	 * @param array  $metadata   Metadata.
-	 *
-	 * @return void
-	 */
-	public function add_pending_webhook( int $order_id, string $new_status, string $old_status, array $metadata = [] ): void {
-		add_post_meta(
-			$order_id,
-			self::PENDING_WEBHOOKS_KEY,
-			[
-				'new_status' => $new_status,
-				'metadata'   => $metadata,
-				'old_status' => $old_status,
-			]
-		);
-	}
-
-	/**
-	 * Get the pending webhooks for an order.
-	 *
-	 * @since 5.18.1
-	 *
-	 * @param int $order_id Order ID.
-	 *
-	 * @return array
-	 */
-	public function get_pending_webhooks( int $order_id ): array {
-		return (array) get_post_meta( $order_id, self::PENDING_WEBHOOKS_KEY );
-	}
-
-	/**
-	 * Delete the pending webhooks for an order.
-	 *
-	 * @since 5.18.1
-	 *
-	 * @param int $order_id Order ID.
-	 *
-	 * @return void
-	 */
-	public function delete_pending_webhooks( int $order_id ): void {
-		delete_post_meta( $order_id, self::PENDING_WEBHOOKS_KEY );
-	}
-
-	/**
 	 * Attempts to get the database option for the valid key from Stripe
 	 * This function was introduced to enable a cache-free polling of the database for the Valid Key, it will include a
 	 * filter to the WordPress All Options and remove the WordPress request cache for the option we are looking at.
@@ -531,26 +481,6 @@ class Webhooks extends Abstract_Webhooks {
 
 		wp_send_json_success( [ 'is_valid_webhook' => false, 'updated' => false, 'status' => $status ] );
 		exit;
-	}
-
-	/**
-	 * Get the max number of retries for the webhooks.
-	 *
-	 * @since 5.19.3
-	 *
-	 * @return int The number of retries.
-	 */
-	public function get_max_number_of_retries(): int {
-		/**
-		 * Filter the maximum number of attempts we will try to retry a webhook process.
-		 *
-		 * @since 5.19.3
-		 *
-		 * @param int $max_attempts How many attempts we will try to retry a webhook process. Defaults to 5.
-		 *
-		 * @return int
-		 */
-		return (int) apply_filters( 'tec_tickets_commerce_gateway_stripe_webhook_maximum_attempts', 5 );
 	}
 
 	/**
